module_29_testing/hw/requirements_test.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pytest==8.3.2
pytest-cov==5.0.0
factory-boy==3.5.0
faker==30.0.0
